feat: extract AeSdkMethods class with minimal interface
